requirements.txt
asgiref==3.9.1
attrs==25.3.0
black==25.1.0
click==8.2.1
Django==5.1.12
django-environ==0.12.0
djangorestframework==3.16.1
drf-spectacular==0.28.0
environ==1.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.4.0
psycopg==3.2.10
psycopg-binary==3.2.10
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.27.1
sqlparse==0.5.3
uritemplate==4.2.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0